OTHER: Improve input validation and group banded functions
Moved validation code to the new module pybaselines._validation and improved input validation. Grouped all banded functions from utils, whittaker, and _algorithm_setup to _banded_utils for easier maintenance.
