+++ pybaselines/utils.py
@@ -11,9 +11,10 @@ from math import ceil
 import numpy as np
 from scipy.ndimage import grey_opening
 from scipy.signal import convolve
-from scipy.sparse import diags, identity
 
+from ._banded_utils import difference_matrix as _difference_matrix
 from ._compat import jit
+from ._validation import _check_scalar
 
 
 # the minimum positive float values such that a + _MIN_FLOAT != a
@@ -410,29 +411,10 @@ def difference_matrix(data_size, diff_order=2, diff_format=None):
     doing penalized least squared fitting or Whittaker-smoothing-based fitting.
 
     """
-    if diff_order < 0:
-        raise ValueError('the differential order must be >= 0')
-    elif data_size < 0:
-        raise ValueError('data size must be >= 0')
-    elif diff_order > data_size:
-        # do not issue warning or exception to maintain parity with np.diff
-        diff_order = data_size
-
-    if diff_order == 0:
-        # faster to directly create identity matrix
-        diff_matrix = identity(data_size, format=diff_format)
-    else:
-        diagonals = np.zeros(2 * diff_order + 1)
-        diagonals[diff_order] = 1
-        for _ in range(diff_order):
-            diagonals = diagonals[:-1] - diagonals[1:]
-
-        diff_matrix = diags(
-            diagonals, np.arange(diff_order + 1),
-            shape=(data_size - diff_order, data_size), format=diff_format
-        )
-
-    return diff_matrix
+    # difference_matrix moved to pybaselines._banded_utils in version 1.0.0 in order to more
+    # easily use it in other modules without creating circular imports; this function
+    # exposes it through pybaselines.utils for backwards compatibility in user code
+    return _difference_matrix(data_size, diff_order=diff_order, diff_format=diff_format)
 
 
 def optimize_window(data, increment=1, max_hits=3, window_tol=1e-6,
@@ -499,61 +481,6 @@ def optimize_window(data, increment=1, max_hits=3, window_tol=1e-6,
         opening = new_opening
 
     return max(half_window, 1)  # ensure half window is at least 1
-
-
-def _check_scalar(data, desired_length, fill_scalar=False, **asarray_kwargs):
-    """
-    Checks if the input is scalar and potentially coerces it to the desired length.
-
-    Only intended for one dimensional data.
-
-    Parameters
-    ----------
-    data : array-like
-        Either a scalar value or an array. Array-like inputs with only 1 item will also
-        be considered scalar.
-    desired_length : int
-        If `data` is an array, `desired_length` is the length the array must have. If `data`
-        is a scalar and `fill_scalar` is True, then `desired_length` is the length of the output.
-    fill_scalar : bool, optional
-        If True and `data` is a scalar, then will output an array with a length of
-        `desired_length`. Default is False, which leaves scalar values unchanged.
-    **asarray_kwargs : dict
-        Additional keyword arguments to pass to :func:`numpy.asarray`.
-
-    Returns
-    -------
-    output : numpy.ndarray
-        The array of values with 0 or 1 dimensions depending on the input parameters.
-    is_scalar : bool
-        True if the input was a scalar value or had a length of 1; otherwise, is False.
-
-    Raises
-    ------
-    ValueError
-        Raised if `data` is not a scalar and its length is not equal to `desired_length`.
-
-    """
-    output = np.asarray(data, **asarray_kwargs)
-    ndim = output.ndim
-    if not ndim:
-        is_scalar = True
-    else:
-        if ndim > 1:  # coerce to 1d shape
-            output = output.reshape(-1)
-        len_output = len(output)
-        if len_output == 1:
-            is_scalar = True
-            output = np.asarray(output[0], **asarray_kwargs)
-        else:
-            is_scalar = False
-
-    if is_scalar and fill_scalar:
-        output = np.full(desired_length, output)
-    elif not is_scalar and len_output != desired_length:
-        raise ValueError(f'desired length was {desired_length} but instead got {len_output}')
-
-    return output, is_scalar
 
 
 def _inverted_sort(sort_order):
